[refactor] revert post fetching methods & widgets into main.dart & deleted posts_screen.dart
